requirements.txt
appdirs==1.4.3
appnope==0.1.0
decorator==4.0.11
Django==1.10.6
django-pure-pagination==0.3.0
ipython==5.3.0
ipython-genutils==0.1.0
markdown2==2.3.3
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
PyMySQL==0.7.10
pyparsing==2.2.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
